- url launcher added to open cv,email,git etc. - Design for My projects added - Model Class added
