+++ ios/Runner/Info.plist
@@ -4,6 +4,10 @@
 <dict>
 	<key>CFBundleDevelopmentRegion</key>
 	<string>$(DEVELOPMENT_LANGUAGE)</string>
+	<key>LSApplicationQueriesSchemes</key>
+    <array>
+      <string>mailto</string>
+    </array>
 	<key>CFBundleDisplayName</key>
 	<string>Naman Portfolio App</string>
 	<key>CFBundleExecutable</key>
