+++ lib/core/constants.dart
@@ -1,5 +1,3 @@
-import 'dart:ui';
-
 import 'package:flutter/material.dart';
 
 const primaryColor = Color(0xFFFF6464);
