+++ lib/features/data/model/MyProjectsModel.dart
@@ -1,0 +1,87 @@
+class MyProjectsModel {
+  final String projectTitle;
+  final String projectDesc;
+  final String projectImage;
+  final String projectLink;
+
+  MyProjectsModel(
+      {required this.projectTitle,
+      required this.projectDesc,
+      required this.projectImage,
+      required this.projectLink});
+}
+
+List<MyProjectsModel> myProjectList = [
+  MyProjectsModel(
+      projectTitle: "Purplle Online Beauty Shopping",
+      projectDesc:
+          "The Purplle app is a very pocket-friendly app with Beauty & Cosmetic products having various top brands such as Mamaearth, Good Vibes, Lakme, Maybelline, L’Oreal, Sugar & many more.",
+      projectImage:
+          "https://lh3.googleusercontent.com/drive-viewer/AITFw-yHgePfqQxDXpvbWDFH65_lodqU5hDmR2pcaQMGE6jfDceUgZEib-FSjORnm-gHVCIf2Am2bkTHJZys__meyoAF5Ntb8w=s1600",
+      projectLink:
+          "https://play.google.com/store/apps/details?id=com.manash.purplle"),
+  MyProjectsModel(
+      projectTitle: "121doc - The Online Clinic App",
+      projectDesc:
+          "Is an online healthcare provider that specialises in a range of prescription treatments for conditions like diabetes and health needs like contraception and period delay. Our secure and discreet service allows you to safely order treatment from our UK-based pharmacy. We're fully licensed and reviewed by the MHRA, CQC, GMC and Data Protection Act.",
+      projectImage:
+          "https://lh3.googleusercontent.com/drive-viewer/AITFw-xgm3SNI47Oy8w-9_O9GGz69ozI3BJYdkI8KYxLNDE7nWWWSvy6qbfNg3-raO9qGuPSE4S-9mAePASMpPMMdOC_1XGu_w=s1600",
+      projectLink:
+          "https://play.google.com/store/apps/details?id=com.hhc.app.artifact_doc121"),
+  MyProjectsModel(
+      projectTitle: "HealthExpress",
+      projectDesc:
+          "HealthExpress has been the online clinic of choice since 2011, providing a convenient, discreet and secure healthcare service. Our patients have can consult our doctors online, request prescriptions and receive treatments from the comfort and privacy of their home. All consultations and prescriptions are carried out by EU-registered doctors and all treatments are shipped from EU-regulated pharmacies.",
+      projectImage:
+          "https://lh3.googleusercontent.com/drive-viewer/AITFw-wJuO0oXZmkMQFLj7BHPU64o7LHkeb9tGh3-T3iTzga7N_FNVPkPQqkTyUWjlY85RMi_XPV4jcSxy_lsDDrh3EQCLKI=s1600",
+      projectLink:
+          "https://play.google.com/store/apps/details?id=com.hhc.app.artifact_healthexpress"),
+  MyProjectsModel(
+      projectTitle: "euroClinix -Your in App clinic",
+      projectDesc:
+          "The online clinic service euroClinix has been providing patients with conclusive information and effective medications since 2011. Using this app, you can create consultations, update your details and reorder treatment with just a couple of swipes.",
+      projectImage:
+          "https://lh3.googleusercontent.com/drive-viewer/AITFw-zVadpuZpIE-AMeOrT79WUMIZYhe81vnX3l3IVr9emPxTIy7m_KKepA3NEihNr8X-9fgCHXW5ypVL1xncgIhcdnQEGk=s1600",
+      projectLink:
+          "https://play.google.com/store/apps/details?id=com.hhc.app.artifact_euroclinix"),
+  MyProjectsModel(
+      projectTitle: "OnlineClinic",
+      projectDesc:
+          "OnlineClinic is one of the UK's first clinics offering prescription treatment online. Working together with our registered UK pharmacy and partnered doctors to dispense medications directly to your door, to date we have helped more than 1,000,000 patients. There are no appointments, opening hours and long wait times. We provide convenient and comprehensive healthcare at the click of a button.",
+      projectImage:
+          "https://lh3.googleusercontent.com/drive-viewer/AITFw-xmertspWtNLOKn65aKSFy8o_E5S8JCV8uvv1P7FLhDNP1hJm7X_xrAwibnHglh_3zGWrqj3dif-_mloBcfVZokfGZE_Q=s1600",
+      projectLink:
+          "https://play.google.com/store/apps/details?id=com.hhc.app.artifact_onlineclinic"),
+  MyProjectsModel(
+      projectTitle: "Ctunes:Calling & Callertune App",
+      projectDesc:
+          "Our app lets you choose what plays on the receiver's phone when you call them - it's the first of its kind in the world. Choose from a variety of media, such as images, videos, audio files, or recorded voices, to make your conversations more memorable and personal.",
+      projectImage:
+          "https://lh3.googleusercontent.com/drive-viewer/AITFw-wLgfbtYNWlqBOXnUgOa5MNM-paSIxcNj19XHT16xrdk9_GW6wnp4ppXW2hfZd8w808R_HLAEEK4NtNWPOw2YwdEGdT=s2560",
+      projectLink:
+          "https://play.google.com/store/apps/details?id=biz.ctunes.callingtunes"),
+  MyProjectsModel(
+      projectTitle: "Indian Post Info",
+      projectDesc:
+          "This Application provides you the list of India's pin code details.You can find post office details by division name, pin code, district name.You can make a search of near by post office depends on your current location on the map and navigate there by selecting current location.",
+      projectImage:
+          "https://lh3.googleusercontent.com/drive-viewer/AITFw-zhlJEo9UXfgZn6CgPOvIT12z0b1FvrtjmbNGMF2OVW23wdGMzJSfD4D8mH1Yt03fLvi984FIEK03V_x4qLShmcxNeeBA=s1600",
+      projectLink:
+          "https://play.google.com/store/apps/details?id=com.rultech.indianpostinfo"),
+  MyProjectsModel(
+      projectTitle: "Reflex Game",
+      projectDesc:
+          "Want to test your memory skills or give your brain an exercise? Try this fun memory game to improve your memory, speed, get better accuracy, This is a fun logical game that gives a good test to your brain.",
+      projectImage:
+          "https://lh3.googleusercontent.com/drive-viewer/AITFw-yiiJghCB1khjUFl9QC861gaMNC6ALscEzt7ayd1TYdmrMZUESxMLy2M8JWLoVzo6RZLCkMm71ietIYes3T7wcYoRJzGQ=s1600",
+      projectLink:
+          "https://play.google.com/store/apps/details?id=rultech.com.reflex_game_flutter"),
+  MyProjectsModel(
+      projectTitle: "Note Keeper - Notepad and ToDo",
+      projectDesc:
+          "Note Keeper is the ultimate productivity companion designed to simplify your life and keep you organized. Whether you're a student, professional, or simply looking to streamline your daily tasks, our Note Keeper app has you covered.",
+      projectImage:
+          "https://lh3.googleusercontent.com/drive-viewer/AITFw-xnpUlSxHanHpr_4HSby-eu0ZQOfiRJmtUaHaoYDz76_Ylth4RB9mXz6ozRYRfnHOSjABzHjr8IRMUG6oo4wmW5OlRk_Q=s1600",
+      projectLink:
+          "https://play.google.com/store/apps/details?id=com.naman_kashyap_flutter_developer.note_keeper_flutter_app"),
+];
