+++ pubspec.yaml
@@ -37,6 +37,7 @@ dependencies:
   cupertino_icons: ^1.0.2
   google_fonts: ^6.0.0
   flutter_svg: ^2.0.7
+  url_launcher: ^6.1.14
 
 
 dev_dependencies:
